refactor: moved all types to the types folder
