--- src/config/types/index.ts
@@ -1,2 +1,0 @@
-export * from './custom-types';
-export * from './types';
